LayoutTests/inspector/styles/styles-add-blank-property.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    WebInspector.showPanel("elements");
    InspectorTest.selectNodeWithId("inspected", step1);

    var treeElement;

    function step1()
    {
        InspectorTest.addResult("Before append:");
        InspectorTest.dumpSelectedElementStyles(true);

        var section = WebInspector.panels.elements.sidebarPanes.styles.sections[0][1];
        section.expand();

        // Create and increment.
        treeElement = section.addNewBlankProperty();
        treeElement.startEditing();
        treeElement.nameElement.textContent = "margin-left";
        treeElement.nameElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));
        treeElement.valueElement.textContent = "1px";
        treeElement.valueElement.firstChild.select();
        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Up"));
        InspectorTest.runAfterPendingDispatches(step2);
    }

    function step2()
    {
        // Increment again.
        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Up"));
        InspectorTest.runAfterPendingDispatches(step3);
    }

    function step3()
    {
        // Commit.
        treeElement.nameElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));
        InspectorTest.selectNodeWithId("other", step4);
    }

    function step4()
    {
        InspectorTest.selectNodeWithId("inspected", step5);
    }

    function step5()
    {
        InspectorTest.addResult("After append:");
        InspectorTest.dumpSelectedElementStyles(true);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a new blank property works.
</p>

<div id="inspected" style="font-size: 12px">Text</div>
<div id="other"></div>

</body>
</html>
